more debugging; fix merging bug
the merging bug is subtle, and I wasn't able to write a unit test
for it...

it happens when we pull in a `seed_dep` via multiple different
ways, so that when we calculate the `base_st`, it is still
included, and subsequently gets applied again, even though we
haven't actually placed it multiple times into any queue.

e.g.
graph with A <- B
seed_deps=[A,B,C]
conc_evid=A
but cur - conc still contains conc, because it gets pulled
in via B, I suppose.